Moved HandleGraphPopulation to own file
